Set the EDITOR/VISUAL to "vim" specifically
This is important with OS X setups so that "/usr/bin/vi" is not used,
EVER (it is an old broken version of vim compared to the ones provided
by Homebrew and MacPorts).

Without this, some features of vim (such as the "+ clipboard register)
do not work, and certain normally inconsequential errors (such as
searching for a string that is not found in the buffer) cause git to
freak out when editing commit messages, refusing to continue the commit
due to a non-zero return code.