profile: add viq alias
This alias is designed for quoting emails with '> ' quoting style, with
a 72-character width limit, reformatting lines as needed to make them
more readable. It automatically imports the clipboard, adds the quotes,
formats the lines, and then exports back to the clipboard, remaining
open for further editing if needed.

This will save me a lot of time when writing emails!