mr-check: OCD script to ensure repos are correct
This script parses the user's .mrconfig file, harvests the list of
"git clone" commands for git repository directories, and then checks
each of those directories' origin fetch remotes to ensure they match.

I found that over time, repositories would move etc. and thus my actual
clones would be out of date or invalid. For example, sometimes I'd have
an old SVN version of a repository that had since moved to git. The mr
program, as cool as it is, doesn't do anything to notify you of this
situation; it simply detects that the folder has an SVN repo and
diligently tries to update, even though if the directory hadn't existed
it would have populated it using a completely different SCM tool...