bashrc: improve git-remote-mediawiki instructions
The executables do not need to go into Git's exec-path; they only
need to go _somewhere_ on your PATH for Git to pick them up.

This is much better because then you do not have to put them in a
(probably) system-wide location.

E.g., on OS X, the git installations are version-specific:

    $ git --exec-path
    /usr/local/Cellar/git/2.1.3/libexec/git-core

Hence, whenever "brew upgrade" installs a new version of Git,
support for git-remote-mediawiki must be manually relinked.

But putting them elsewhere on the path does not have that issue.