bashrc: first cut at modularization
This splits the monolithic bashrc configuration into a few different
shell script plugins, oh-my-zsh style.

The ultimate goal is to make each of these plugins completely
standalone, so that none depend on any of the others (because none of
the shell frameworks I found explicitly support plugin dependency
management). But at the moment, a couple of the plugins require others
to be loaded first, according to the current lexicographic ordering.