Use "stub" rather than "template" suffix
These stub files are not templates (intended for copy-paste-edit),
but rather placeholders that add a level of indirection.

This is useful for systems that do not support symlinks, so that when
the dotfiles Git repository is updated, the configuration still updates
automatically. (I.e., the stub files essentially never change, and
hence can be copied rather than linked.)