vimrc: use 'vundle' not 'Vundle.vim' as directory
It bothers me a little to do this, because 'Vundle.vim' is the
official repository name now, and 'vundle' is the old name, which
only still works thanks to GitHub's awesome renaming support.

The reason to use 'vundle', though, is because the oh-my-zsh
vundle plugin (which is very nice) still uses the old name.