Use built-in function for diff
Because diff was aliased to git, zsh tries to be clever and use
git's tab completion logic for diff arguments. Unfortunately, for
whatever reason, file paths do not autocomplete in this scenario.

Using a built-in function is suitably indirect, so that the default
tab completion logic is used, and file paths autocomplete as desired.