zshrc: _really_ disable automatic updates
From glancing at the source, it seems the DISABLE_AUTO_UPDATE variable
does not actually disable the prompt, but rather disables the update
prompt? Furthermore, in practice, I'm still seeing the prompt even with
this variable set.

A more surefire approach is to set the LAST_EPOCH variable in
~/.zsh-update to an absurdly large value, so that the update check
always terminates without prompt.