config-links.sh: fix nonstandard test
At this point, I have no idea where I got "-a" as a test flag.
Its only standard use is for "and"-style concatenation.

What we really want to do here is remove symlinks, and rename regular
files. So writing "else if the file is a regular file..." seems fine.