java.sh: improve JAVA_HOME detection on Linux
This leans on the update-java-alternatives program on Linux to discover
Java 1.6, 1.7 and 1.8 and set them into the J6, J7 and J8 environment
variables.

It also provides a handy jhome built-in for DRYness which lets you ask
for any Java version (not just 1.6, 1.7, and 1.8).

Unfortunately, the j6, j7 and j8 aliases still don't fully work on Linux
because /usr/bin/java does not respect the value of JAVA_HOME. You have
to run "update-java-alternatives -s" as root to change the Java version
system-wide, which is beyond the scope of these configuration files.

An alternative would be to put $JAVA_HOME/bin into the path, and/or
alias the relevant java* commands to $JAVA_HOME/bin/java* instead.
Maybe later.