Add script to search <developers> & <contributors>
This is handy when filling in the developers and contributors
sections of new projects, so that the blocks match existing ones.

The included directories to search are currently hardcoded,
which works fine for me, but would require generalization to make
it more generally useful for a broader audience. Hence, I include
the script here in ctr-scripts rather than scijava/scijava-scripts.