mr-check: warn about uninitialized local dirs
As a control freak, I like to know when my mr-config wants a directory
to be a Git repository, but I don't actually have it cloned.

Without this knowledge, I might do e.g. "mr up" in a folder expecting
to have a couple of repositories get updated, and then in addition,
some other folders end up getting initialized/cloned!