bashrc: remove redundant debug output
Each plugin already reports itself that it's loading
(because with zsh it's a single command to load all).

And fix plugins to report their names correctly with bash.
Using $0 won't work with bash here, unfortunately, because
the bashrc uses "source plugins/foo.sh" which apparently
doesn't result in a new set of $ argument variables.