Switch from zgen to zpm
The zgen project has not been maintained for a few years.
There are a couple of forks, but nothing looked promising to me.
The zpm project, on the other hand, is active and pretty nice.

This commit also adds a few nice plugins I discovered as part of the
migration, and removes obsolete code covered better by those plugins.