Always use absolute link paths + update mr config
There is just too much room for error with ../.. and so forth.
Using absolute paths frees us from the need to think about how
many levels deep we are, and avoids the need for case logic
depending on the location of the dotfiles.

This change also moves the mrconfig.d linkages from ~/.mrconfig.d to
~/.config/mr, and links in only essential by default. Users should link
in desired repository groups manually on a machine-by-machine basis.

Finally, various mr repository groups have been updated.