Remove mercurial repos and git-remote-hg
The bisque and bisquik repositories do not appear to be
accessible from the ucsb.edu URLs anymore.

I found Bisque on GitHub at:
    https://github.com/UCSB-VRL/bisque

But the last update to the mainline branch was in 2019.

The newer project looks to be Bisque 2, available from:
    https://github.com/UCSB-VRL/bisqueUCSB

which is being actively updated. But when I cloned it,
it was 745MB. Since I'm not actively working with it,
let's hold off on tracking it here in these dotfiles.