install-packages: do not install vim-gtk anymore
It was renamed to vim-gtk3 in Ubuntu 22.10.

And I'm too lazy to add case logic to this script right now.
Later, the script should probably check every single package on the list
for existence with `test apt-cache show $p >/dev/null 2>&1` or similar,
and only install the packages that exist.

Open question whether we need to ensure anything like "install vim-gtk
XOR vim-gtk3" (i.e. not both), for some theoretical situation where both
packages exist in the package manager but one of them is considered more
desirable than the other. To handle that situation, we may need to
introduce a syntax in the packages lists like `vim-gtk3>vim-gtk` where
it tries each one and chains through the `>` symbols when non-existent.