Only source bashrc from bash_profile if interactive
This may avoid naive additions to ~/.bashrc, notably `export DEBUG=1`,
from "infecting" non-bash shell instances throughout the login session.

But note that this change is *not sufficient* on my system to avoid
~/.bashrc ultimately getting sourced by non-interactive login shells.
See previous commit for a change that actually fixes the issue.

Why not? Because Ubuntu Linux's stock ~/.profile contains this:

    # if running bash
    if [ -n "$BASH_VERSION" ]; then
        # include .bashrc if it exists
        if [ -f "$HOME/.bashrc" ]; then
            . "$HOME/.bashrc"
        fi
    fi

While we could override the stock ~/.profile here in dotfiles, there
is a question of whether the override would still be appropriate on
FreeBSD-based systems such as macOS. So I'm not going to do that.